refactor: separate out components
